libmyth: Avoid a divide by 0 exception in AudioOutputBase
If MythMusic starts when the default audio is ALSA and if the device
is in use by another process then snd_pcm_open can fail.  In this
case AudioOutputBase::SetAudiotime can cause a divide by 0 exception
because this->effdsp is 0.

This patch checks the value of effdsp and avoids the division if it's 0.

Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>