libmyth: Avoid a divide by 0 exception in AudioOutputBase
If MythMusic starts when the default audio is ALSA and if the device
is in use by another process then snd_pcm_open can fail.  In this
case AudioOutputBase::SetAudiotime can cause a divide by 0 exception
because this->effdsp is 0.

This patch checks the value of effdsp and avoids the division if it's 0.

Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>
(cherry picked from commit 84aac9d80f98a30d3cb68caf4827e9e92caedbc1)