Move QUERY_FILETRANSFER handling to improve backend performance.
This significantly reduces the backend CPU usage for me when streaming a
recording or other file to the frontend.

Each time we send a command to the backend we perform a number of string
comparisons to find the correct handler. In the cases where that same command
is sent regularly, maybe several times a second those comparisons can stack up
to result in a high CPU load. This can be avoided by moving the most frequently
used commands to the top of the list thereby avoiding a couple of dozen or more
comparisons each time.

There may be other commands that should be moved to the top with similar
benefit.
(cherry picked from commit 529853d11587ebee0b77753272d6dcc1a5404a11)