Convert LoggingItem to use ReferenceCounter
This gets rid of the deleteLater() which removes the nasty memory growth
issues in various threads that don't use the full Qt event loop.

Also converted mthreadpool processEvents/sendPostedEvents to use qApp->
rather than QCoreApplication::  QCoreApplication:: is specifically for console
apps, which made mythfrontend grow like crazy as it did not effect it.