Move the queue locking to a much smaller scope in LogPrintLine
We don't need to lock the queue to create the items, but only when handling
the queue.  This should make future debugging of potential deadlocks simpler,
and also make logging slightly faster in the case that two threads are trying
to "simultaneously" log something.  The only thing that needs to be serialized
is the access to the queue itself.  The rest can be in parallel.