use qIsNaN in place of ::isnan
Feature patch for g++ C++0x/C++11 standards compliance

When compiling with g++ and specifing -std=c++0x the compilation fails in
mythplayer.cpp due to a namespace collision for isnan (std:: and math::).

Patch attached which changes mythplayer to use the Qt function qIsNaN,
providing a platform neutral implementation now that Qt has made that
function available (in Qt 4.6).

Since MythTV does not currently specify use of the new C++ standards, this
patch is not necessary at this time, but may useful be at some future date
(when gcc defaults to the C++11 standards?)

Fixes #10524

Signed-off-by: Gavin Hurlbut <ghurlbut@mythtv.org>