Remove use of OF zlib macro in MythGame.
The macro seems to only be of any use for compilers that are not ANSI-C
compliant, which should be... nothing these days. Chances are the rest
of the code base wouldn't compile on it anyway. On anything modern, the
macro doesn't do anything. Remove it and see what happens. Comments in
zconf.h seem to indicate it may cause problems for OSX.

Refs #10507