Convert the MythSystemPrivate classes to use RefCounters
We were getting some odd race conditions caused by the timing of deletion of
the MythSystemPrivate data (MythSystemUnix in particular on Linux and OSX).
As we now have a common ReferenceCounter class to use, I have converted this
over to use ReferenceCounter to control the deletion.  One reference is made
on creation, and a second when it is added to the process map.  The second
reference is decremented when the MythSystemSignalThread is done shutting it
down cleanly.

NOTE:  I may have missed some dependency on the parent item which may have
been deleted already.  If this happens, please let me know.  It would most
likely crash hard in this case.  I'll take another look over it in a bit.

Refs #10775.  Refs #10889.