Attempt to plug a memory leak in MPEG2 "lossless" transcode
Seems that when we hit the 2000 frames in use (i.e. queued), we very
ungracefully reject more frames, "No more queue slots!", and then we promptly
leak the packet we were going to put in that frame.  Added a few more
av_free_packet() to hopefully plug the leak.

Also, since it seems the HDHR Prime (and perhaps other sources) record MPEG2
that can sometimes have exceedingly large GOP sizes.  This code was mostly
designed for use with ivtv and others with predicably small GOP sizes.  To
try to get the "No more queue slots!" to disappear, I have increased the
max frames from 2000 to 20000.  This could still use a rework, but hopefully
will fix the short-term issue.

Thanks to rkulagow for reporting this mess. :)