Allows multiple sessions to Airplay and RAOP
This is done by stopping the one currently running or stopping any current playback.

This change allows to always be able to start airplay and not silently fail should a playback be currently going. If a current playback is currently going (either RAOP, AirPlay or standard playback), it will be stopped and a new session will be automatically started.
AirPlay now waits synchronously for any events to be completed: this result in much smoother user interaction and also follow Apple's user experience guidelines with their MFi vendors.

A side effect is that since iOS 5.1, a typical AirPlay session would open both a RAOP session and an AirPlay one, to drop the RAOP connection a few seconds later. This caused the audio card to be already in use for AirPlay and resulted in no audio during playback

There is an issue with iOS devices when using RAOP, they don't detect that the connection got interrupted and continue sending data. iTunes behaves properly.

Fixes #10885