Add a weather grabber using the official MetOffice API for the UK
This will probably replace the broken BBC grabbers.

This is still a work in progress. Currently offers 6 (just 5 available) and 3
day screens. An 18 hour and possibly a new 24 hour screen will be
added soon. A current observations grabber is also planned but a lower
priority for now.

Search doesn't include any caching yet, but the plan is to have it
fully cached.

Please report bugs to the mailing list.