Add playback registration handling
It is an unfortunate fact that you can only do one playback at a time within a frontend; usually because you can only open one audio session at a time.
This commit provides a mechanism for playback objects (such as TV, AirPlay, RAOP, MythMusic) to register themselves as a player and request playback; in which case all other registered player will be asked to close.