HLS encoder enhancements.
- Use AAC audio instead of MP3.  This negates the need for
using --enable-libmp3lame on the command line since we now use the
built in AAC encoder in libavcodec.  This change also requires
converting the audio data to float for use by the encoder.  For
now, this commit uses C code to do the int to float conversion,
but a call to use AudioOutputUtil::toFloat() to use assembler to
do this conversion should be enabled in the near future.

- Use High 4.1 profile for encoding video with a width of >= 1024.
I also think the changes made for this support might have fixed an
issue with parts of the main profile not being used previously
for output video >= 960 in width.