Remove Painter settings on Mac OS X
MythFrontend requires OpenGL for proper operation, and mythtv-setup requires Qt.
This is a quick and dirty way to remove the guesswork for the user and automatically have what works.

MythMainWindow::Init now can take an optional argument specifying which theme painter to use and do not rely on user configuration

Fixes #10469