Fixes #10926. Safely teardown MythPainters.
The deadlock was actually the least of our problems.
Since MythPainter::DeleteFormatImagePriv() is a pure
virtual in MythPainter calling ExpireImages() in
the dtor is unsafe. This puts the teardown code
into a Teardown() method which is called from all
the child classes where it is safe to do the teardown.