Fix for Synaesthesia memory corruption
We were allocating just half the required space and consequently
writing beyond the end of the array. This consistently caused
segfaults for me when using the Synaesthesia visualiser.

This fix appears to be consistent with the code used in other projects
which were also derived from the original Synaesthia project.

e.g. GStreamer/Source/gst-plugins-ugly/gst/synaesthesia/synaescope.c