Fixes #10872. Use desired start and end times.
The recording quality measurement needs to know when a recording was
supposed to start and end to know how much is actually missing from
the recording. We used to use the scheduled start and end times for
this, but sometimes a recording is intentionally started late or ended
early. This captures that by tracking the 'desired' start and end point.
This is initially set by the scheduler using and pre and post roll
values, and if the user intentionally stops a recording early that
updates the desired end time as well.

Conflicts:

	mythtv/libs/libmythtv/dtvrecorder.h