Make SignalHandler a singleton.
There can only be one SignalHandler object per application.
This allows AddHandler to be a static method so it can
be used outside the main initialization. Since AddHandler
doesn't actually chain handlers as suggested by it's name
this also renames the method to SetHandler.

This also adds a lock for m_sigMap so we can call SetHandler
in a thread other than the main thread.

Finally, this allows a handler to be added for SIGINT and
SIGKILL.