Fixes #10970. Fix StreamHandler restart race.
We wait for _running to be set to false, but we need the thread to be
in the finish state or QThread::start() will exit silently without
starting the thread. Since we're calling QThread::wait() we don't
really need the while (_running) loop at all in either Start() or
Stop(), so I've removed it. This race probably dates back to the
pthread->QThread porting.
(cherry picked from commit 7d01e691d2bb58ed7dcbbdcca93ec06dc3a7381b)