MediaMon: Ensure to send status changed after disc mounted
If a USB key disc is inserted then a mount event is never triggered
and the media is never recognized.  If the disk is mounted before
mythfrontend is started or is mounted by an external agent then the
media is recognized.

This patch ensures that whenever a disk is mounted an event is triggered
and also adds some debug log output to show all media transitions.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>