Revert [6fb8692] to not use aacenc.c AAC encoder by default.
We've been recommended not to use the native AAC encoder by default
since there are still some known issues with the encoder.  This
commit tries to be as minimally invasive as possible since we're
close to release, but it does change it so that the native AAC
encoder in libavcodec is not used unless neither libmp3lame nor
libfaac support was configured.

If the native AAC encoder is working fine for you then you don't
have to do anything.  If you want to use libmp3lame as MythTV v0.25
used, then configure MythTV with --enable-libmp3lame.  If you wish
to use AAC but prefer libfaac over the native AAC encoder, then
you can run configure with --enable-libfaac --enable-nonfree.  If
both --enable-libmp3lame and --enable-libfaac are used, then
libfaac will take precedence.