Mythgallery: Fix white screen after playing video when using OpenGL
If the mythfrontend UI is set to use OpenGL and MythGallery is set to use
OpenGL transitions then when video playback exits a white screen is
shown instead of the video screenshot with "Press SELECT to play again"
overlay.

The fix is to call makeCurrent on return from playback.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>
(cherry picked from commit c0907083ab091a42222548312bb3b43f59feeb82)