Fix UI message interface availability when using MythWelcome.
This adds a mechanism to disable the UI message interface when running
MythFrontend, freeing up the necessary ports to allow the frontend to
listen on them and receive events while MythWelcome idles in the
background. The interface is brought back up on MythWelcome when the
frontend terminates.

This commit contains an ABI bump for two new events and a MythSystem
constant.

Fixes #10815