Complete timezone handling in Python bindings following UTC change.
This adds proper timezone handling to the Python bindings, parsing zone
information out of the system TZif files. On systems without those
tables available, the code will revert to a simple offset, yielding
potentially inaccurate times when used on old recordings.

All user interaction with datetimes will be in local time, unless forced
otherwise by a timezone-aware datetime object. Times are converted to
UTC automatically for interfacing with MySQL and mythbackend.

Refs #10941

I'm leaving this ticket open for now, as the timezones will still need
to be created in a different manner on Windows, and potentially OSX.