Reinitialize database session variables when reconnecting.
The session variables weren't being set when we reconnected after a
database connection went stale and led to problems.  For example, the
timezone was not being set to UTC and caused serious scheduling
problems for users in far eastern time zones.  Thanks to Adrian Saul
for initially spotting and then tracking down the problem.

Fixes #11086