Add better handling of simple zoneinfo files.
This adds support for timezones with no defined transitions (UTC), and
ones that define a single transition for a simple offset. This also
improves the normal search function to properly handle dates that fall
to the future of the last transition point, and thus cannot be
bracketed. Dates in the past before the first transition point give a
more descriptive error.

Fixes #11093