Calculate current playback position using frame count and frame rate.
According to Taylor Ralph, the code which uses timestamps is not
finished and shouldn't be used yet.  For one thing, it doesn't handle
discontinuities in the timestamps.

Note: timestamps are still used sometimes to calculate recording
lengths.  That will need to be fixed in a future commit.

(cherry picked from commit 286186281dcca4a3f25f591e6658310bc9d375fb)