Fix video encode in MPEG2 "lossless" transcoding
Fixes #11118

It seems that the new video encode API call pipelines the encoding, returning
after each input frame, but not returning a packet the first time around in all
cases.  This change makes it encode the packet, and then keep calling the API
until a packet has actually been encoded (flushing the output).