Add reporting of logging volume to Smolt.
This reports the logging volume as a percentage of the total amount
stored in the database, for each urgency level. This is intended as a
simple metric to determine stability of MythTV, as higher counts of
higher urgency levels will indicate an increase in problems seen by the
user base.

(cherry picked from commit 5f90b1d62c4d692a6c23ae22ab2047db85e6572e)