mythtranscode code re-org
Move the following classes out of transcode.cpp to make things a little
cleaner and allow later reuse of the audio and video classes in a
HTTP Live Streaming encoder which will not be contained within transcode.cpp.

  AudioBuffer
  AudioReencodeBuffer
  Cutter
  TranscodeFrameQueue (renamed to VideoDecodeBuffer)

There is no new functionality in this patch, only a code re-org.