Fix immutable methods in Artwork class.
This fixes an issue with several immutable methods in the Artwork class,
inherited from the parent UserString class. As the Artwork class is
modified for use as a property, and not intended to be called directly
for storage of a string, the inherited UserString methods that returned
a new instance using the __class__ attribute would fail. This adds a
__new__ method that detects such uses and forces the call to return a
standard str type instead.