Move SendEvent/SendMessage processing out of UI thread.
When I re-enabled the SendReceiveStringList UI blocking debugging it turned out
almost all the UI blocking for MythProto was caused by processing asynchronous
events via a synchronous MythProto call in the UI thread. This creates a runable
to process these outside the UI thread so we get a much smoother UI experience.