mythmusic: Fix streaming radio not being able to connect to some streams
This fixes a couple of errors with creating the stream GET request header.
If the path part of the the URI is empty the HTTP/1.1 spec says it should be
specified as "/" and not be empty and the Host request-header field should
contain the port if it is non standard.

Fixes #11214

Signed-off-by: Stuart Morgan <smorgan@mythtv.org>
(cherry picked from commit fe22adc5c485b8b19969b0e1804a84f652227577)