mythmusic: Save playlists to DB when changed not at frontend exit
This changes things so any changes to playlists are saved to the database
immediately rather than cached and saved when the frontend exits.

This fixes a problem were you could lose edits to a playlist if the FE is not
shutdown cleanly like when using the idle timer to shutdown or if you shutdown
the computer from the FE exit menu.
(cherry picked from commit ed2097986db80798ca00694542a9cb9d0e61c855)

Fixes #11228

Signed-off-by: Stuart Morgan <smorgan@mythtv.org>