Remove support for recording rule type priorities.
This change removes support for the arcane recording rule type
priorities that probably nobody or almost nobody actually uses.  This
is the feature where recording rules are given a priority boost based
on their type (e.g. SingleRecord, ChannelRecord, AllRecord, etc.) in
addition to the per-rule and other guide based priorities.  If anyone
actually uses this feature, it can be replicate by using the almost as
obscure custom priorities feature.

Note: this is the first in a series of changes over the coming days
and weeks to remove some scheduler related features and settings that
the developers no longer believe are needed or want to support.