Remove the SchedMoveHigher scheduler setting.
This setting controls whether or not the scheduler can move a high
priority program that is already scheduled to another time, channel or
input to make room for a low priority program.  The scheduler now
always does this.  In the rare case the resulting schedule is
undesirable, overrides can be used to achieve the desired result.