Fix AirTunes so it works with iTunes 11
iTunes used to only send an AES-128 challenge once, and myth was expecting that the challenge came during the OPTIONS packet.
The challenge is now sent with more packet, including during the ANNOUNCE one.
So we move the handling of the challenge so it will work no matter which packet it comes with.

Tested with iTunes 11, iTunes 10 and iOS 6.01