Remove and restrict the use of some recording rule types.
This is another change in the series of changes to remove some
scheduler related features and settings that the developers no longer
believe are needed or want to support.

Remove the Timeslot and Weekslot recording rule types.  These rule
types are too rigid and don't work when a broadcaster shifts the
starting time of a program by a few minutes.  Users should now use
Channel recording rules in place of Timeslot and Weekslot rules.  To
approximate the old functionality, two new schedule filters have been
added.  In addition, the new "This time" and "This day and time"
filters are less strict and match any program starting within 10
minutes of the recording rule time.

Restrict the use of the FindDaily and FindWeekly recording rule types
(now simply called Daily and Weekly) to search and manual recording
rules.  These rule types are rarely needed and limiting their use to
the most powerful cases simplifies the user interface for the more
common cases.  Users should now use Daily and Weekly, custom search
rules in place of FindDaily and FindWeekly rules.

Any existing recording rules using the no longer supported or allowed
types are automatically converted to the suggested alternatives.