Don't leak nextRecording or nextRingBuffer.
If a recorder was deleted after SetNextRecording has been called it
would leak a RecordingInfo and a RingBuffer. The RecordingInfo is
no big deal but the RingBuffer is holding an open file descriptor
and the application will start to misbehave if after we leak just
a few hundred of those since select() will no longer function
correctly.