Suppress clang++'s unused variable warnings.
clang++ is too smart for it's own good here. It detects that ZMQ_ASSERT's
aren't using their values in the profile build. Other than these it doesn't
detect any additional unused values over g++ and icpc so we suppress
these in the clang build.

Note: We do not suppress the "unused-result" warnings which do
catch bugs the other compilers do not..