Remove redundant "All Programs - " text in Watch Recordings screen.
Fixes #11027.  Removes the "DispRecGroupAsAllProg" setting (Show
filter name instead of "All Programs"), and always uses "All Programs"
as the first group name.  Removes the unnecessary prefix from the
titles of the group entries.

Fixes the (not-yet documented) "group" widget of the watchrecordings
window in recordings-ui.xml to always hold the current group filter
name.  Themers are suggested to display the current group filter in a
textarea, e.g.:

  <value>Watch Recordings</value>
  <template>Watch Recordings - %1</template>