Remove the LiveTVPriority setting.
This is another change in the series of changes to remove some
scheduler related features and settings that the developers no longer
believe are needed or want to support.

The scheduler now always tries to avoid interrupting live TV.
Recordings that conflict with live TV are moved to other inputs or
later times as needed.  Live TV is only interrupted if a recording
can not be moved.