Select the "best" video tracks available for playback
Optimize and speed-up the detection and handling of video streams. Every video streams used to be opened during the scan, and the last one found would be use.
This commit change the behavior as follow:
Scan all streams.
Once scan is completed, search for the best video stream using FFmpeg convenience method av_find_best_stream. And only open the code and initialise the video output for that particular stream.

Also, should an error occur while scanning a particular stream, it is now simply ignored instead of completely stopping the scan.