Partially resolve unicode filename issue in ftopen().
This resolves a unicode character issue when trying to use the ftopen()
helper function to access files. The ftopen() function accepts a
standard myth:// URI, and parses it for the necessary information,
however URIs traditionally only accept a limited, safe subset of the
ASCII character set. This adds a temporary work around where by the
various object .open() methods will internally pass the necessary data
in a tuple, bypassing the need to encode and decode about the limited
character set. Myth URIs with unsafe or escaped characters will still
not be handled properly by ftopen().