Add some more scheduler related simplifications.
Remove the channel record recording rule type.  Add a "This channel"
filter to maintain the that functionality.  Automatically convert
existing channel record rules to all record rules with the channel
filter enabled.

Add a dialog to offer common actions instead of immediately entering
the schedule editor when SELECT is pressed on a non-scheduled program.
Pressing the EDIT action still goes straight to the schedule editor.
Change the text in the recording, not recording and not scheduled
dialogs to be a little clearer.

Change the behavior when RECORD is pressed in the EPG and elsewhere.
Previously, this action cycled through the recrding rule types.  if
the program is not scheduled, this action now adds a single record
rule.  If the program is scheduled, it now brings up the appropriate
recording or not recording dialog.