Fix an issue with keyframe positions produced by mythcommflag --rebuild.
The seeking code in DecoderBase assumes there is a keyframe at frame
0, otherwise it shifts all keyframe numbers by the value of the first
keyframe number in the map.  This causes problems for recordings with
variable keyframe distances.  To correct this, make sure mythcommflag
--rebuild creates an explicit entry for frame 0.

Note: this does not fix the issue of mythcommflag --rebuild adding
bogus keyframes to the seektable for h.264 recordings.

Cherry-picked from a small part of
49dbed5be0729b04a5f0fd0426a32fceb2dd7935 .