Bump MDM::downloadNow() timeout up to 60 seconds.
The previous 10-second timeout was causing issues when trying to
download large lineups from Schedules Direct. This commit bumps
the downloadNow() timeout to 60 seconds.  This is a blocking call,
so if the caller doesn't want to potentially block for that long,
they should be use one of the non-blocking methods.

Fixes #11287.