Fix an issue with total duration in files without seektables.
For recordings/videos without seektables, the decoder dynamically adds
keyframes to the in-memory table as it encounters them.  We don't want
to do that for the duration map since each entry represents the
cumulative time of all prior frames, and this gets messed up if
playback doesn't start at the beginning, or jumps forward or backward
at all.  The exception is mythcommflag --rebuild, which plays every
frame from start to end.