MythArchive: fix the missing logs when doing a native archive import/export
The new logging stuff was causing both the export and import of native archives
logging to not be shown in the log viewer this fixes that and also makes the
log viewer wait a while for the logs before giving up which should fix the
logs not always showing the first time.

Fixes #10505.
(cherry picked from commit 52a7c118aedc747bbea3a92d178eabf8566818ad)