Enable 'best match' detection for recording metadata lookups.
This enables the automatic mode which attempts to identify the correct
match when multiple results are returned. This is already used for
videos but was for some reason disabled for recordings.

This may result in incorrect metadata for some recordings and we'll
try to fix that as best we can, but it should generally behave better
than before where a high percentage of lookups resulted in no artwork
at all because most lookups matching multiple results resulted in failure.