Add support for IMDB based lookups
I'm giving greg credit for this, but the method is significantly
differnet. This includes an explicit regular expression check to ensure
only zero-padded seven digit numbers are queried as IMDB IDs. This is to
prevent any risk of potential overlap with TMDB IDs, once their database
grows that large. This also does not bother applying the IMDB IDs to the
temporary VideoMetadata object as MythTV does not attempt to do anything
with it.

Fixes #11345

Signed-off-by: Raymond Wagner <rwagner@mythtv.org>