Improve the Manual Zoom experience during video playback.
Add new ZOOM* actions to the TV Playback context (with no predefined
key bindings), to get away from overloading other key bindings.  The
old overloads are left intact, but the ZOOM actions are preferred.

Double the maximum zoom in/out factors.

Remove the behavior of wrapping back to 1.0 when trying to zoom in or
out beyond the maximum factor.