Migrate Screen Setup Wizard to external application.
This sidesteps the issue of trying to display the corner arrows outside
the current extents of the UI by starting up an entirely new instance
and forcing it to full screen. The arrows are additionally replaced by
the theme's preview image, to more clearly exhibit the final appearance.

There are some issues and potential usability concerns that still need
to be addressed.
* The preview image repositions, but does not properly scale.
* There may need to be indicators to show what corner is currently used.
* The network control interfaces are not available.

This changeset includes a revision bump, and will require any theme with
a custom appear-ui.xml to be updated.